ProductServiceImpl.java

+++ E-Commerce/src/main/java/com/springproject/ecommerce/service/impl/ProductServiceImpl.java
@@ -50,9 +50,7 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public List<Product> findProductByCategoryId(Integer categoryId) {
-		
-		
+	public List<Product> findProductByCategoryId(Long categoryId) {
 		
 		return prodRepo.findProductByCategoryId(categoryId);
 	}
